fix: adding back in bitwise operator checks since apparently they are needed
