+++ packages/eslint-plugin/tests/rules/strict-enums-assignment.test.ts
@@ -13,6 +13,33 @@ const valid: ValidTestCase<unknown[]>[] = [];
 const invalid: InvalidTestCase<MessageIds, unknown[]>[] = [];
 
 valid.push({
+  name: 'Declaring an enum with an empty initializer',
+  code:
+    fruitEnumDefinition +
+    `
+  let fruit: Fruit;
+  if (true) {
+    fruit = Fruit.Apple;
+  } else {
+    fruit = Fruit.Banana;
+  }
+    `,
+});
+
+/**
+ * In development, this would trigger run-time errors due to the
+ * `typeChecker.getTypeAtLocation` method being buggy and not having a proper
+ * function signature.
+ */
+valid.push({
+  name: 'Declaring an "empty" variable with array destructuring',
+  code: `
+  const myArray = [1];
+  const [firstElement] = myArray;
+    `,
+});
+
+valid.push({
   name: 'Assigning a number enum literal to a number enum (with type-inference)',
   code:
     fruitEnumDefinition +
@@ -244,6 +271,15 @@ valid.push({
     fruitEnumDefinition +
     `
 const fruitArray: Fruit[] = [];
+  `,
+});
+
+valid.push({
+  name: 'Assigning a new variable with a composition of bitflags',
+  code:
+    fruitEnumDefinition +
+    `
+const fruitFlags = Fruit.Apple | Fruit.Banana;
   `,
 });
